Correções e Implementação de ModelMapper

+++ src/main/java/com/lojaVirtual/lojaVirtual/services/produto/interfaces/MarcaService.java
@@ -15,5 +15,5 @@ public interface MarcaService {
 
     boolean criarMarca(MarcaDTO marca);
 
-    boolean deletaMarca(int id);
+    boolean deletaMarca(long id);
 }
