+++ src/main/java/com/lojaVirtual/lojaVirtual/services/produto/interfaces/ProdutoService.java
@@ -14,5 +14,5 @@ public interface ProdutoService {
 
     boolean criarProduto(ProdutoDTO categoria);
 
-    boolean deletaProduto(int id);
+    boolean deletaProduto(long id);
 }
